+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Changelog
  * removed puli/composer-plugin dependency
  * upgraded webmozart/path-util to 2.0
  * renamed `puli asset` to `puli publish`
+ * renamed `puli binding` to `puli bind`
 
 * 1.0.0-beta4 (2015-04-13)